# backend-order/docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateOrderRequest:
    properties:
      customer_id:
        type: string
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - customer_id
    - product_id
    - quantity
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      payment_id:
        type: string
      product:
        $ref: '#/definitions/models.OrderProduct'
      quantity:
        type: integer
      status:
        type: string
      timeline:
        items:
          $ref: '#/definitions/models.TimelineEvent'
        type: array
      total_amount:
        type: number
      updated_at:
        type: string
    type: object
  models.OrderProduct:
    properties:
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  models.Product:
    properties:
      id:
        type: string
      name:
        type: string
      price:
        type: number
      stocks:
        type: integer
    type: object
  models.TimelineEvent:
    properties:
      name:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a simple backend server using Go and Gin framework.
  title: Order API
  version: "1.0"
paths:
  /admin/orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders from all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all orders
      tags:
      - admin
      - orders
...
  /orders:
    get:
      description: Get the list of orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get my orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order (requires authentication)
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/api.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an existing order (requires authentication)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Cancel an order
      tags:
      - Orders
...
  /products/{id}:
    get:
      description: Get detailed information of a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product by ID
      tags:
      - Products
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"